Add TAsyncTransportBase to declare async interface for transports
